refactor(creator-shop): consume shared useSpotlight hook in callout
Extract the cursor-following spotlight idiom (already copy-pasted ~10x, incl.
a divergent inline hook in the Referrals dashboards) into a shared
`~/hooks/useSpotlight`, parametrized by radius + gradient color, and have the
Cosmetic Studio callout consume it instead of a hand-rolled copy. Migrating
the remaining call sites is a separate cleanup.
